yangsongflask/requirements.txt
ansi2html==1.8.0
black==23.11.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
future==0.18.3
idna==3.4
importlib-metadata==6.8.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
platformdirs==4.0.0
plotly==5.18.0
protobuf==4.21.12
PyMySQL==1.1.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
